Got "equation_evaluator" logic almost working in MathJS_UnitsTesting2.html
This does everything from the python equation_evaluator.py file.. The only difference is it converts into other 'simplified' units. For example:
59751781.2789155 s⁻¹ ÷ 1×10⁶ = 59.7517812789155 MHz
However, this is not a big issue because JSONGrapher will make the same units for everything in the end. That being said, I need to modify the function to take the units *each* iteration so that each datapoint can be converted into common units.